Allow user color selections to be enforced
Previously if the theme didn't define a color palette for the editor there were inconsistencies between the editor and the frontend. Color selections made by the user were different on the frontend if relying only on the named css classes. This is an attempt to force that via syncing inline styles.